Prepared a source parser for behavioral computer (#7)
This source parser is a sub-type of the FileChecker, which recursively searches through a file tree. However, this Checker only searches through sub-directories of the top level "source" directory based on the contents of an indicator dir, which contains indicator files. Using a regex and a format string, the name of each file in the indicator dir is converted into a folder name to search. If there is a matching folder in the main source dir, it is searched for new files recursively. Otherwise it is ignored.

Commits:
* Added a IndicatorFileChecker

* Added better error handling when bad indicator files passed

* Fix the recursion behavior with inheritance

* Refactor clean to better support code re-use, include clean for indicator

* Refactor upload stat to its own directory, made middle location optional

* Refactor prepare.py to use the state_path

* Fix some minor key and formatting issues

* Supress divide by zero warning when no files transfered